# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blog Project API",
        "description": "API documentation for the blog system.",
        "contact": {
            "email": "farzane.rahmani2@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/profile/image": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Upload user's profile image",
                "operationId": "3c26c51013a0c61a85adea32a30d8e6d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image"
                                ],
                                "properties": {
                                    "image": {
                                        "description": "Image file to upload",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Image uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Profile image updated successfully."
                                                },
                                                "profile_image_url": {
                                                    "type": "string",
                                                    "example": "http://localhost:8000/storage/profile_images/vl05yIorRvteG9ROPKBkYlbfP5vP851c0nwuF3hf.jpg"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "server_time": {
                                            "type": "string",
                                            "example": "2025-05-24 16:51:01"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/rankings": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get list of top ranked users by total post views",
                "operationId": "09bb9de4d4cb6c8983077bc4c14435b4",
                "responses": {
                    "200": {
                        "description": "Successful response with ranked users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "users": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Tara Pagac"
                                                            },
                                                            "profile_image": {
                                                                "type": "string",
                                                                "example": "http://localhost:8000/storage/profile_images/ukzz2WRY2zouwJd72zFJJzLNBoLyww1jKDkxawll.jpg"
                                                            },
                                                            "total_views": {
                                                                "type": "integer",
                                                                "example": 4
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "server_time": {
                                            "type": "string",
                                            "example": "2025-05-24 17:31:01"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Unexpected server error"
                    }
                }
            }
        },
        "/api/profile/me": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get the currently authenticated user's profile",
                "operationId": "a0825aa0966fdc2d7c3fe0ea9ad215b7",
                "responses": {
                    "200": {
                        "description": "User profile retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 26
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "farzane"
                                                    },
                                                    "mobile": {
                                                        "type": "string",
                                                        "example": "09197224528"
                                                    },
                                                    "profile_image": {
                                                        "type": "string",
                                                        "example": "http://localhost:8000/storage/profile_images/ZbApL64uXQ5LNkhqA9C3ZIyQwmVTn2EAe5KMKafy.jpg"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "server_time": {
                                            "type": "string",
                                            "example": "2025-05-24 17:18:46"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}